@@ -1,9 +1,14 @@
 module gitoa.ru/go-4devs/encoding
 
-go 1.16
+go 1.22
 
 require (
-	github.com/BurntSushi/toml v0.3.1
-	github.com/stretchr/testify v1.7.0
-	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b
+	github.com/BurntSushi/toml v1.6.0
+	github.com/stretchr/testify v1.11.1
+	gopkg.in/yaml.v3 v3.0.1
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
 )
