@@ -5,7 +5,7 @@ import (
 )
 
 // Encode writer from volume.
-type Encode func(w io.Writer, v interface{}) error
+type Encode func(w io.Writer, v any) error
 
 // Decode reader to volume.
-type Decode func(r io.Reader, v interface{}) error
+type Decode func(r io.Reader, v any) error
