@@ -12,10 +12,10 @@ var (
 	_ encoding.Encode = Encode
 )
 
-func Decode(r io.Reader, v interface{}) error {
+func Decode(r io.Reader, v any) error {
 	return json.NewDecoder(r).Decode(v)
 }
 
-func Encode(w io.Writer, v interface{}) error {
+func Encode(w io.Writer, v any) error {
 	return json.NewEncoder(w).Encode(v)
 }
