@@ -1,7 +1,7 @@
 package encoding
 
 // Unmarshal bytes to inteface.
-type Unmarshal func(data []byte, v interface{}) error
+type Unmarshal func(data []byte, v any) error
 
 // Marshal interface to bytes.
-type Marshal func(v interface{}) ([]byte, error)
+type Marshal func(v any) ([]byte, error)
